+++ packages/backend/src/altNodes/altConversion.ts
@@ -1,5 +1,3 @@
-import { convertNodesOnRectangle } from "./convertNodesOnRectangle";
-
 type StyledTextSegmentSubset = Omit<
   StyledTextSegment,
   "listSpacing" | "paragraphIndent" | "paragraphSpacing" | "textStyleOverrides"
@@ -31,6 +29,24 @@ export const cloneNode = <T extends BaseNode>(node: T): T => {
   }
 
   return cloned;
+};
+
+/**
+ * Identify all nodes that are inside Rectangles and transform those Rectangles into Frames containing those nodes.
+ */
+export const convertNodesOnRectangle = (
+  node: FrameNode | GroupNode | InstanceNode | ComponentNode | ComponentSetNode,
+): FrameNode | GroupNode | InstanceNode | ComponentNode | ComponentSetNode => {
+  if (node.children.length < 2) {
+    return node;
+  }
+  if (!node.id) {
+    throw new Error(
+      "Node is missing an id! This error should only happen in tests.",
+    );
+  }
+
+  return node;
 };
 
 export const frameNodeTo = (
